remove provider
