+++ aksel.nav.no/website/pages/templates/theme-switch/uten-system.tsx
@@ -1,6 +1,6 @@
 import { ThemeIcon } from "@navikt/aksel-icons";
 import { Button } from "@navikt/ds-react";
-import { useThemeExample } from "@/web/examples/ThemeExample.context";
+import { useThemeExample } from "@/web/examples/useThemeExample";
 import { withDsExample } from "@/web/examples/withDsExample";
 
 function Example() {
